Trait positioning::pathfinding::Heuristic

pub trait Heuristic {
    fn heuristic_distance(
        &self,
        start: Position,
        end: Position
    ) -> WithInfinity<u64>;

    fn find_shortest_path(
        &self,
        open_positions: &BTreeSet<Position>,
        start: Position,
        end: Position
    ) -> Option<VecDeque<Position>> { ... }
}

An admissible heuristic for the A* pathfinding algorithm is one which always returns an optimistic result. Oftentimes, a good heuristic can make a big performance difference. The more tight the heuristic is to the real distance, the better.

Required Methods

fn heuristic_distance(&self, start: Position, end: Position) -> WithInfinity<u64>

Provided Methods

fn find_shortest_path(
&self,
open_positions: &BTreeSet<Position>,
start: Position,
end: Position
) -> Option<VecDeque<Position>>